+++ src/Service/QueryParametersParser.php
@@ -5,10 +5,6 @@ namespace Lmc\ApiFilter\Service;
 use Lmc\ApiFilter\Assertion;
 use Lmc\ApiFilter\Entity\Value;
 use Lmc\ApiFilter\Exception\TupleException;
-use Lmc\ApiFilter\Exception\UnknownFilterException;
-use Lmc\ApiFilter\Filter\FilterIn;
-use Lmc\ApiFilter\Filter\FilterInterface;
-use Lmc\ApiFilter\Filter\FilterWithOperator;
 use Lmc\ApiFilter\Filters\Filters;
 use Lmc\ApiFilter\Filters\FiltersInterface;
 use MF\Collection\Exception\TupleExceptionInterface;
@@ -18,6 +14,14 @@ use MF\Collection\Immutable\Tuple;
 
 class QueryParametersParser
 {
+    /** @var FilterFactory */
+    private $filterFactory;
+
+    public function __construct(FilterFactory $filterFactory)
+    {
+        $this->filterFactory = $filterFactory;
+    }
+
     public function parse(array $queryParameters): FiltersInterface
     {
         try {
@@ -38,7 +42,7 @@ class QueryParametersParser
             })
                 ->reduce(
                     function (FiltersInterface $filters, ITuple $tuple): FiltersInterface {
-                        return $filters->addFilter($this->createFilter(...$tuple));
+                        return $filters->addFilter($this->filterFactory->createFilter(...$tuple));
                     },
                     new Filters()
                 );
@@ -71,25 +75,5 @@ class QueryParametersParser
         return $columnsCount > 1
             ? Tuple::parse($value, $columnsCount)->toArray()
             : [$value];
-    }
-
-    private function createFilter(string $column, string $filter, Value $value): FilterInterface
-    {
-        switch (mb_strtolower($filter)) {
-            case 'eq':
-                return new FilterWithOperator($column, $value, '=', 'eq');
-            case 'gt':
-                return new FilterWithOperator($column, $value, '>', 'gt');
-            case 'lt':
-                return new FilterWithOperator($column, $value, '<', 'lt');
-            case 'lte':
-                return new FilterWithOperator($column, $value, '<=', 'lt');
-            case 'gte':
-                return new FilterWithOperator($column, $value, '>=', 'gte');
-            case 'in':
-                return new FilterIn($column, $value);
-        }
-
-        throw UnknownFilterException::forFilterWithColumnAndValue($filter, $column, $value);
     }
 }